refactor: extract bill editor keyboard height logic
